Implement review suggestions by @csadorf
